Extract PowerHelper
